admin

+++ routes/web.php
@@ -1,10 +1,7 @@
 <?php
 
-
-use App\Http\Controllers\TamuController;
 use App\Http\Controllers\PegawaiController;
 use App\Http\Controllers\ProfileController;
-
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\AdminController;
 
@@ -23,27 +20,20 @@ Route::get('/edit-pegawai/{x}', [PegawaiController::class, 'edit'])->middleware(
 Route::post('/update-pegawai/{x}', [PegawaiController::class, 'update'])->middleware(['auth', 'verified'])->name('pegawai.update');
 Route::delete('/hapus-pegawai/{id}', [PegawaiController::class, 'hapus'])->middleware(['auth', 'verified'])->name('pegawai.hapus');
 
-
-
-
-Route::get('/tamu', [TamuController::class, 'index'])->middleware(['auth', 'verified'])->name('tamu');
-Route::get('/tambah-tamu', [TamuController::class, 'tambah'])->middleware(['auth', 'verified'])->name('tamu.tambah');
-Route::post('/simpan-tamu', [TamuController::class, 'simpan'])->middleware(['auth', 'verified'])->name('tamu.simpan');
-
+Route::get('/tamu', function () {
+    return view('tamu');
+})->middleware(['auth', 'verified'])->name('tamu');
 
 Route::get('/user', function () {
     return view('user');
 });
 Route::get('/admins', [AdminController::class, 'index'])->middleware(['auth', 'verified'])->name('admins');
-Route::get('/admins', [adminController::class, 'create'])->middleware(['auth', 'verified'])->name('admins');
-Route::post('/admins', [adminController::class, 'edit'])->middleware(['auth', 'verified'])->name('edit');
+Route::get('/admins/create', [adminController::class, 'create'])->middleware(['auth', 'verified'])->name('admins.create');
+
 Route::post('/admins/{x}', [adminController::class, 'update'])->middleware(['auth', 'verified'])->name('update');
-Route::delete('/admins.hapus/{id}', [adminController::class, 'hapus'])->middleware(['auth', 'verified'])->name('hapus');
+Route::delete('/admins.hapus/{id}', [adminController::class, 'hapus'])->middleware(['auth', 'verified'])->name('admins.hapus');
 
-    
-
-
-Route::resource('/simpan-tamu_user', \App\Http\Controllers\Tamu_UserController::class);
+Route::post('/admins/edit/{id}', [AdminController::class, 'edit'])->name('admins.edit');
 
 Route::middleware('auth')->group(function () {
     Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
